Restructured labeling of Spring components.
What previously had been called Component is now an AnnotatedInjectable. This allows to introduce Component as separate concept to capture classes annotated with @Component. AnnotatedInjectables are labeled with Injectable now. Extended the rules for allowed component stereotype interdependencies to include generic Components as well.

Tweaked the integration tests to remove explicit checks for the generic Injectable concept if they actually check for more concrete concepts like Repository or a Configuration.